Move `API\Model\*transaction*` to `API\Model\Payments`
