+++ includes/api/clients/blockchain/class-btc-rpc-explorer-api.php
@@ -12,7 +12,7 @@ namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Clients\Blockchain;
 
 use BrianHenryIE\WP_Bitcoin_Gateway\API\Blockchain_API_Interface;
 use BrianHenryIE\WP_Bitcoin_Gateway\API\Clients\Blockchain\Adapters\Btc_Rpc_Explorer_Api_Transaction_Adapter;
-use BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Transaction_Interface;
+use BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Payments\Transaction_Interface;
 use BrianHenryIE\WP_Bitcoin_Gateway\BtcRpcExplorer\BtcRpcExplorerApi;
 use BrianHenryIE\WP_Bitcoin_Gateway\BtcRpcExplorer\Model\TXSummary;
 use Psr\Log\LoggerAwareInterface;
