+++ includes/api/model/payments/class-transaction-vin.php
@@ -9,7 +9,7 @@
  * @package brianhenryie/bh-wp-bitcoin-gateway
  */
 
-namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model;
+namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Payments;
 
 use BrianHenryIE\WP_Bitcoin_Gateway\Brick\Money\Money;
