+++ includes/api/model/payments/interface-transaction-interface.php
@@ -6,7 +6,7 @@
  * @package brianhenryie/bh-wp-bitcoin-gateway
  */
 
-namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model;
+namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Payments;
 
 use DateTimeInterface;
