+++ includes/api/repositories/queries/class-bitcoin-transaction-query.php
@@ -8,7 +8,7 @@
 namespace BrianHenryIE\WP_Bitcoin_Gateway\API\Repositories\Queries;
 
 use BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Payments\Bitcoin_Transaction_WP_Post_Interface;
-use BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Transaction;
+use BrianHenryIE\WP_Bitcoin_Gateway\API\Model\Payments\Transaction;
 use BrianHenryIE\WP_Bitcoin_Gateway\WP_Includes\Post_BH_Bitcoin_Transaction;
 use DateTimeInterface;
